requirements.txt
numpy
pandas
opencv-python
scikit-image
plotly
pyyaml
pytest